LG-4876 rename withProps.jsx
