+++ app/javascript/packages/document-capture/components/document-capture.jsx
@@ -14,7 +14,7 @@ import { BackgroundEncryptedUploadError } from '../higher-order/with-background-
 import SuspenseErrorBoundary from './suspense-error-boundary';
 import SubmissionInterstitial from './submission-interstitial';
 import PromptOnNavigate from './prompt-on-navigate';
-import withProps from '../higher-order/withProps';
+import withProps from '../higher-order/with-props';
 
 /** @typedef {import('react').ReactNode} ReactNode */
 /** @typedef {import('./form-steps').FormStep} FormStep */
