+++ app/javascript/packages/document-capture/higher-order/with-props.jsx
